reformatting register groups with proper bitmapping
