+++ Libraries/LTC6811_2/option_enums.h
@@ -1,24 +1,24 @@
-
-
-//TODO: change to enum casses instead?
-//not sure if it would really be any more beneficial
-
-namespace LTC6811_2 {
+#pragma once
+#include <string.h>
+#include <stdint.h>
+#ifdef HT_DEBUG_EN
+    #include "Arduino.h"
+#endif
 
 //datasheet defined choices
-enum ADC_MODE {
+enum class ADC_MODE {
 	MODE_ZERO=0,
 	FAST,
 	NORMAL,
 	FILTERED
 };
 
-enum DISCHARGE {
+enum class DISCHARGE {
 	DISCHARGE_NOT_PERMITTED = 0,
 	DISCHARGE_PERMITTED
 };
 
-enum CELL_SELECT {
+enum class CELL_SELECT {
 	ALL_CELLS = 0,
 	CELL_1_7,
 	CELL_2_8,
@@ -28,17 +28,17 @@ enum CELL_SELECT {
 	CELL_6_12
 };
 
-enum OPEN_WIRE_CURRENT_PUP {
+enum class OPEN_WIRE_CURRENT_PUP {
 	PULL_DOWN=0,
 	PULL_UP
 };
 
-enum SELF_TEST_MODE{
+enum class SELF_TEST_MODE{
 	SELF_TEST_1=1,
 	SELF_TEST_2
 };
 
-enum GPIO_SELECT {
+enum class GPIO_SELECT {
 	ALL_GPIO_AND_REF_2=0,
 	GPIO_1,
 	GPIO_2,
@@ -48,7 +48,7 @@ enum GPIO_SELECT {
 	REF_2
 };
 
-enum STATUS_GROUP_SELECT {
+enum class STATUS_GROUP_SELECT {
 	ALL_GROUPS=0,
 	SC,
 	ITMP,
@@ -56,32 +56,7 @@ enum STATUS_GROUP_SELECT {
 	VD
 };
 
-
-//better name later
-enum CELL_DISCHARGE {
-	CELL_1 = 1<<0,
-	CELL_2 = 1<<1,
-	CELL_3 = 1<<2,
-	CELL_4 = 1<<3,
-	CELL_5 = 1<<4,
-	CELL_6 = 1<<5,
-	CELL_7 = 1<<6,
-	CELL_8 = 1<<7,
-	CELL_9 = 1<<8,
-	CELL_10 = 1<<9,
-	CELL_11 = 1<<10,
-	CELL_12 = 1<<11
-};
-
-enum GPIO_CONFIG {
-	GPIO1_SET = 1<<0,
-	GPIO2_SET = 1<<1,
-	GPIO3_SET = 1<<2,
-	GPIO4_SET = 1<<3,
-	GPIO5_SET = 1<<4,
-};
-
-enum S_CONTROL {
+enum class S_CONTROL {
 	S_DRIVE_HIGH = 0,
 	S_1_PULSE,
 	S_2_PULSES,
@@ -91,5 +66,4 @@ enum S_CONTROL {
 	S_6_PULSES,
 	S_7_PULSES,
 	S_DRIVE_LOW
-};
 };
